spi: move schema to spi
