+++ core/schema/src/test/java/org/eclipse/dataspaceconnector/schema/SchemaTest.java
@@ -14,6 +14,8 @@
 
 package org.eclipse.dataspaceconnector.schema;
 
+import org.eclipse.dataspaceconnector.spi.types.domain.schema.Schema;
+import org.eclipse.dataspaceconnector.spi.types.domain.schema.SchemaAttribute;
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.Test;
