+++ spi/src/main/java/org/eclipse/dataspaceconnector/spi/types/domain/schema/Schema.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.dataspaceconnector.schema;
+package org.eclipse.dataspaceconnector.spi.types.domain.schema;
 
 import java.util.AbstractSet;
 import java.util.LinkedHashSet;
